Skip deleting when user is Administrator

+++ tieta/tieta/doctype/cell_station/cell_station.py
@@ -34,9 +34,8 @@ class CellStation(Document):
 		self.save()
 
 	def on_trash(self):
-		self.set("site", None)
-		self.save()
-		frappe.delete_doc("Cloud Project Site", self.site, ignore_permissions=True)
+		if frappe.session.user != 'Administrator':
+			frappe.delete_doc("Cloud Project Site", self.site, ignore_permissions=True)
 
 	def on_update(self):
 		site = frappe.get_doc("Cloud Project Site", self.site)
